Fix: 찜 Controller 분리, Redirect 수정 및 UI 수정 #63
